+++ src/Options/EmphasisStyle.cs
@@ -1,0 +1,8 @@
+﻿namespace MarkdownEditor
+{
+    public enum EmphasisStyle
+    {
+        Asterisk,
+        Underscore,
+    }
+}