+++ src/Options/Options.cs
@@ -5,6 +5,7 @@ namespace MarkdownEditor
 {
     public class Options : DialogPage
     {
+        // General
         [Category("General")]
         [DisplayName("Enable outlining")]
         [Description("Determines if outlining (code folding) should be enabled for multiline HTML and code blocks")]
@@ -29,6 +30,20 @@ namespace MarkdownEditor
         [DefaultValue(true)]
         public bool EnableSmartIndent { get; set; } = true;
 
+        // Style
+        [Category("Style")]
+        [DisplayName("Bold character")]
+        [Description("Determines if bold should use double asterisk or underscore. Example: **bold text**")]
+        [DefaultValue(EmphasisStyle.Asterisk)]
+        public EmphasisStyle BoldStyle { get; set; } = EmphasisStyle.Asterisk;
+
+        [Category("Style")]
+        [DisplayName("Italic character")]
+        [Description("Determines if italic should use single asterisk or underscore. Example: _italic text_")]
+        [DefaultValue(EmphasisStyle.Underscore)]
+        public EmphasisStyle ItalicStyle { get; set; } = EmphasisStyle.Underscore;
+
+        // Preview window
         [Category("Preview Window")]
         [DisplayName("Enable Preview Window")]
         [Description("Determines if the preview window should be shown")]
@@ -63,6 +78,7 @@ namespace MarkdownEditor
         [Browsable(false)]
         public double PreviewWindowHeight { get; set; } = 500;
 
+        // Validation
         [Category("Validation")]
         [DisplayName("Enabled")]
         [Description("Enable validation to run on local link and image references.")]